start nvim as server listening pipe file
The pipe file is private to each CWD from where nvim is started.

This should allow to use nvim as the default editor for interactive git
commands by exporting the GIT_EDITOR environment variable, but nvim is
not yet supporting the --remote-wait CLI flag which is required
(otherwise git doesn't wait for the editing to be completed).

Anyway being able to open a file within the currently running nvim
instance seems a nice feature to have, so I'm leaving it on.