synch accepts relative paths during mapping
It seems that nvim changed the way it can pass the file path to the
BufWritePost event; before it was the absolute path, now it can
sometimes be a relative path to the root folder.

The remote synch should therefore detect that and restore the full path
before attempting the search for a path mapping in order to find it,
otherwise the mapping will fail and the file won't be synchronized.