Refactor QuantConnectClient class out of remote_sync feature
This gives the possibility to interact with QuantConnect API also
outside the remote_sync code, for example to setup a custom keymap to
trigger actions like backtesting.

Contextually add functions to trigger project build and backtesting on
the remote.

Correctly implement class structure of both quantconnect manager
(in remote_sync) and client (in extensions)