basic persistent undo history
WARNING: this will only work on new buffers. If a buffer is loaded from
a previous stored session it will need to be unloaded and loaded again
for the persistent history to be applied to it. This is because
undodir/undofile settings are buffer local and the previous session
didn't include those options for the stored buffers.

Things to improve:
 - the diff from undotree plugin is very basic, ideally it should look
   like the diff from Diffview plugin for git
 - add a mapping in the undotree view to show the available mappings in
   that view (usually with g?)
 - add a mapping in the undotree to scroll the diff (similar to
   Diffview)