add BasicIdeInitProject [template]
The command creates a new project configuration file in the home
folder. If the template is specified, it is used to index the defined
templates in the settings to fill the otherwise empty settings table
with some initial values. Normally one would set custom predefined
templates in the user-level config file in the home folder, otherwise
just rely on the default ones from the IDE.

If the requested template doesn't exist an error is reported and the
command does nothing.

If a project config file already exists it is just opened for edit.