avoid spawning infinite nvim processes
in case the alias is not used then the environment variable
SET_TITLE_SCRIPT doesn't exist. If the title starts with "nvim", setting
the title will result in a call to nvim again, which will try to set the
same title, entering an infinite loop.

This will result in many nvim processes being spawned, eventually
killing the machine.